+++ test/unit/tests/routes/utils.unit.ts
@@ -8,14 +8,13 @@ import fixtures from '../../fixtures/utils.fixtures.js';
 describe('utils', () => {
   fixtures.map(fixture => {
     test(fixture.name, async () => {
-      const fastify = buildFastify({ maxParamLength: 32_768 });
-
-      await fastify.ready();
-
       vi.spyOn(config, 'getConfig').mockReturnValue({
         ...config.mainConfig,
         network: fixture.network,
       });
+      const fastify = buildFastify({ maxParamLength: 32_768 });
+
+      await fastify.ready();
 
       const response = await supertest(fastify.server).get(fixture.endpoint);